+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+// Place your settings in this file to overwrite default and user settings.
+{
+  "files.exclude": {
+    "**/.git": true,
+    "**/.DS_Store": true,
+    "build": true,
+    "dist": true,
+    "**/*.js":  { "when": "$(basename).ts" }
+  }
+}
