--- scripts/components/footer.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const React = require('react')
-
-const Footer = (props) => {
-  return React.createElement('footer', { id: 'footer' },
-    React.createElement('div', null, props.database)) }
-
-module.exports = Footer
