--- scripts/components/tree/tree-constants.js
@@ -1,19 +1,0 @@
-'use strict'
-
-const TREE_EXPAND                   = 'TREE_EXPAND'
-const TREE_COLLAPSE                 = 'TREE_COLLAPSE'
-const TREE_FETCH_CHILDREN_FAILURE   = 'TREE_FETCH_CHILDREN_FAILURE'
-const TREE_FETCH_CHILDREN_RESPONSE  = 'TREE_FETCH_CHILDREN_RESPONSE'
-const TREE_FETCH_ITEM_FAILURE       = 'TREE_FETCH_ITEM_FAILURE'
-const TREE_FETCH_ITEM_RESPONSE      = 'TREE_FETCH_ITEM_RESPONSE'
-const TREE_SELECT             = 'TREE_SELECT'
-
-module.exports = {
-  TREE_COLLAPSE,
-  TREE_EXPAND,
-  TREE_FETCH_CHILDREN_FAILURE,
-  TREE_FETCH_CHILDREN_RESPONSE,
-  TREE_FETCH_ITEM_FAILURE,
-  TREE_FETCH_ITEM_RESPONSE,
-  TREE_SELECT
-}
