+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "jsx": "react",
+    "outDir": "build",
+    "target": "es6"
+  },
+  "exclude": [
+    "dist",
+    "node_modules"
+  ]
+}
